feat: add HasTypeIDWithAmount interface
