+++ src/Contracts/IPriceable.php
@@ -27,13 +27,8 @@ namespace Seat\Services\Contracts;
  * This interface is in the services package to encourage making classes that describe items compatible across both the
  * seat core and plugin, even if they don't depend on recursivetree/seat-prices-core.
  */
-interface IPriceable extends HasTypeID
+interface IPriceable extends HasTypeID, HasTypeIDWithAmount
 {
-    /**
-     * @return int The amount of items to be appraised by a price provider
-     */
-    public function getAmount(): int;
-
     /**
      * Set the price of this object.
      *
